refactor(genconfig): extract main to cmd using genconfig as package
